image filters working but not displaying previews

+++ app/src/main/java/com/chenhuiyeh/imagefilter/Interfaces/EditImageFragmentListener.java
@@ -1,7 +1,7 @@
 package com.chenhuiyeh.imagefilter.Interfaces;
 
 public interface EditImageFragmentListener {
-    void onBrightnessChanged(float val);
+    void onBrightnessChanged(int val);
     void onSaturationChanged(float val);
     void onContrastChanged(float val);
     void onEditStarted();
